extract paginated results out into new component; move applicable tests to new test file for said component; ensure search component refactoring is reflected and covered in tests
