+++ ui/src/__tests__/components/Checkbox.test.tsx
@@ -1,5 +1,3 @@
-import '@testing-library/jest-dom';
-
 import * as Components from '@/components';
 
 import { render, screen } from '@testing-library/react';
