+++ ui/src/lib/interfaces.ts
@@ -230,7 +230,7 @@ export interface User extends CoreUser {
     works: WorksRecord[];
 }
 
-export interface SearchResults<T extends Publication | PublicationVersion | User> {
+export interface SearchResults<T extends FlagByUser | Publication | PublicationVersion | User> {
     data: T[];
     metadata: SearchResultMeta;
 }
